+++ src/Plutarch/Extra/List.hs
@@ -23,9 +23,54 @@ module Plutarch.Extra.List (
     pfind',
     pfindJust,
     plookupAssoc,
+
+    -- * Unhoisted helpers
+    psingletonUnhoisted,
+    plist,
+    pmatchList,
 ) where
 
 import qualified Plutarch.List as PList
+
+{- | As 'psingleton', but unhoisted.
+
+ @since 3.4.0
+-}
+psingletonUnhoisted ::
+    forall (ell :: (S -> Type) -> S -> Type) (a :: S -> Type) (s :: S).
+    (PElemConstraint ell a, PListLike ell) =>
+    Term s a ->
+    Term s (ell a)
+psingletonUnhoisted x = plist [x]
+
+{- | Make a list-like structure from any 'Foldable' of 'Term's. This is
+ unhoisted.
+
+ @since 3.4.0
+-}
+plist ::
+    forall (ell :: (S -> Type) -> S -> Type) (f :: Type -> Type) (a :: S -> Type) (s :: S).
+    (Foldable f, PListLike ell, PElemConstraint ell a) =>
+    f (Term s a) ->
+    Term s (ell a)
+plist = foldr (\x xs -> pcons # x # xs) pnil
+
+{- | An \'uncons\' operation, designed for CPS chaining, for arbitrary list-like
+ structures.
+
+ @since 3.4.0
+-}
+pmatchList ::
+    forall (r :: S -> Type) (a :: S -> Type) (ell :: (S -> Type) -> S -> Type) (s :: S).
+    (PElemConstraint ell a, PListLike ell) =>
+    -- | Continuation for \'nil\'.
+    Term s r ->
+    -- | Continuation for \'cons\'.
+    (Term s a -> Term s (ell a) -> Term s r) ->
+    -- | The list-like to match on.
+    Term s (ell a) ->
+    Term s r
+pmatchList = flip pelimList
 
 {- | Returns 'PTrue' if the list-like structure contains at least one element.